feat: refactor, split tests
